move core into hooks/provider, and rename hooks to ts
