+++ packages/convex-helpers/package.json
@@ -24,10 +24,6 @@
       "types": "./dist/react/sessions.d.ts",
       "default": "./dist/react/sessions.js"
     },
-    "./react/cache/core": {
-      "types": "./dist/react/cache/core.d.ts",
-      "default": "./dist/react/cache/core.js"
-    },
     "./react/cache/hooks": {
       "types": "./dist/react/cache/hooks.d.ts",
       "default": "./dist/react/cache/hooks.js"
